+++ example/with-selector/index.tsx
@@ -2,7 +2,7 @@ import React, { FC, memo } from 'react';
 import { createTrackedSelector } from 'react-tracked';
 import { defineModule } from '../../src';
 
-const CounterModule = defineModule({ count: 0, input: 'hello' })
+const counterModule = defineModule({ count: 0, input: 'hello' })
   .actions({
     add: (draft) => draft.count++,
     setInput: (draft, value: string) => (draft.input = value),
@@ -10,7 +10,7 @@ const CounterModule = defineModule({ count: 0, input: 'hello' })
   .build();
 
 export const WithSelector: FC = memo(() => {
-  const [{ count, input }] = CounterModule.use();
+  const [{ count, input }] = counterModule.use();
 
   return (
     <div>
@@ -25,14 +25,14 @@ export const WithSelector: FC = memo(() => {
 });
 
 const InputComponent: FC = memo(() => {
-  const [input, { setInput }] = CounterModule.use((state) => state.input);
+  const [input, { setInput }] = counterModule.use((state) => state.input);
   console.log('input component rerender');
   return <input value={input} onChange={(e) => setInput(e.target.value)} />;
 });
 
-const useCounterModuleState = createTrackedSelector(CounterModule.useState);
+const useCounterModuleState = createTrackedSelector(counterModule.useState);
 const CounterComponent: FC = memo(() => {
-  const { add } = CounterModule.useActions();
+  const { add } = counterModule.useActions();
   const { count } = useCounterModuleState();
   console.log('counter component rerender');
   return <button onClick={add}>{count}</button>;