+++ test/module-factory.test.tsx
@@ -5,7 +5,7 @@ import { defineModule } from '../src';
 describe('test module factory', function () {
   afterEach(cleanup);
 
-  it('factory chain is immutable', function () {
+  it('should factory builder methods be pure function', function () {
     const factory1 = defineModule({ state: 0 });
     const factory2 = factory1.actions({});
 
@@ -13,14 +13,14 @@ describe('test module factory', function () {
       hello: () => {},
     }));
 
-    const Module2 = factory2.build();
-    const Module3 = factory3.build();
+    const module2 = factory2.build();
+    const module3 = factory3.build();
     {
-      const { result } = renderHook(() => Module2.useActions());
+      const { result } = renderHook(() => module2.useActions());
       expect(Object.keys(result.current)).toHaveLength(1); // setState
     }
     {
-      const { result } = renderHook(() => Module3.useActions());
+      const { result } = renderHook(() => module3.useActions());
       expect(Object.keys(result.current)).toHaveLength(2); // setState + hello
     }
   });